Branch Push for Server
Pushing code as is to test git setup on linux server
